ui/base: Clean up our AcceleratorGtk class.
Currently this class is fully implemented in the header file, and this have annoyed me for a long time.
Most of our class implementations reside in source files, so this patch fix this now. And this also makes
it easier to read.

R=sky@chromium.org

Review URL: https://chromiumcodereview.appspot.com/9965049

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@130236 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9294cbb0e98567abb0a2fd6789845a6575951641