Revert 135593 - Add initial GTK web accessibility framework (new with valgrind fix).
This is the same as http://codereview.chromium.org/9839069/ which was
landed but reverted due to a valgrind failure in content_unittests.

The valgrind failure was because browser_accessibility_manager_unittest.cc
is constructing its own subclass of BrowserAccessibility (for testing),
but BrowserAccessibilityManagerGtk was assuming that any BrowserAccessibility
could be static_casted to a BrowserAccessibilityGtk. I fixed this by
making ToBrowserAccessibilityGtk() a virtual method with a default impl that
returns NULL.

Original change:

This enables Linux desktop assistive technology such as the Orca screen
reader to access the web contents. Builds on the same accessibility stack
used by Mac & Windows already.

This change works with Orca now, but it's minimal: all that works is getting
the correct feedback when you tab through focusable links and controls in
a webpage. Future changes will add the rest of the support.

BUG=24585
TEST=Run Chrome on desktop Linux with Orca turned on, tab through links.
TBR=darin,sky
Review URL: https://chromiumcodereview.appspot.com/10381010

TBR=dmazzoni@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10379029

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@135595 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 95b3f544eae3f1de91d29d0f4e1abef05592bd86