Allow the RWHVW to be focused when a window is restored from minimized.
This restores a hack that was in place in the old NativeTabContentsContainer code that seemed to clear focus during restoration (though the hack was for another purpose, the hack seemed to fix this bug too). I've isolated the hack to the re-activation code.

http://crbug.com/125976
TEST=see bug
Review URL: https://chromiumcodereview.appspot.com/10378043

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@136124 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ef291978cb15b511e4b024376fc65b12b89c24fd