Compute the enclosing rectangle for circles/ellipses generated from a series of touch events
For a sequence of {touchstart, touchmove*, touchend} events, compute the rectangle that encloses all of the circles/ellipses corresponding to the touch events. If the hardware is capable of collecting radius_x or radius_y, use the maximum of these two values as the radii values for the touch region. Otherwise use a default radius value of 15.

BUG=122190
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10262023

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@137753 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 748bd3ad873b96f0295f29f71f386dc587c3845c