Draw dragged file text with NO_SUBPIXEL_RENDERING flag.
This is needed because the destination is a transparent canvas.

Note: It's not needed to set that flag for DrawStringWithHalo()
because DrawStringWithHalo() draws text to an intermediate
opaque canvas.

BUG=105550
TEST=Download a file on Aura. Drag the file from the download bar.
The file name text should not look garbled.

Review URL: https://chromiumcodereview.appspot.com/10441046

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@139083 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4065db53278eb558d980f948b7987c01adcaf206