Merge scalable-image grd files using new support in Grit (r51).
{theme,ui}_resources_{standard,2x,touch,touch_1x,touch_2x}.grd are merged into {theme,ui}_resources_standard.grd.

The only visible change should be that theme_resources_2x used to point to the 100% version of settings_favicon.png, but now picks up the 200% version. I checked the 200% image and it looks fine.

Note that this changes the resource IDs because of a grit bug related to processing of <if> nodes.

BUG=131216
TEST=Existing tests, visual inspection, pak files unchanged except for settings_favicon.png and different resource IDs

Review URL: https://chromiumcodereview.appspot.com/10562030

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@142766 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c7b2fa3f761dfa9652e3528387758d5e508f2dea