Android requires some extra functionality to set the path to it's resource pak files and native libraries.
This patch adds a new value to the PathService to implement the resource pak override and update the places that load pak files to take advantage of it.

Also adds a function for embedders to set the path to native libraries.

Review URL: https://chromiumcodereview.appspot.com/10802065

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@149842 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f152b67b5802543986e37a9f57e228e424786e84