Move SessionStorageNamespace entirely into NavigationController and support StoragePartitions.
Previously, WebContents, NavigationController, and RenderViewHost all exposed
APIs that allowed one to retrieve the SessionStorageNamespace associated with
them. This is confusing because there were too many ways to access the object.
After this change, the SessionStorageNamespace is only exposed by the
NavigationController.

Conceptually a SessionStorageNamespace belongs to a WebContents but we store
it in NavigationController because on tab duplication, the
NavigationController becomes the authoritative state of the tab.

Also, to support StoragePartitions, the NavigationController
now maintains a map of partition_id -> SessionStorageNamespace. Someone
requesting a SessionStorageNamespace must either know which StoragePartition
they are coming from, or which child process they are acting on behalf of.

This change also changes the way TabContents and WebContents are created.
1) We now have explicitly separate creation methods for prepopulating
   with SessionStorage and creating with an opener.
2) Some of the WebContentImpl construct has been moved into an Init()
   function to avoid accidental calls to virtual functions by subobjects.

TBR-ing all the directories where I just remove a NULL.

TBR=sky,stevenjb,dominich,brettw,satorux,kalman
BUG=85121

Review URL: https://chromiumcodereview.appspot.com/10831116

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@151379 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d1198fdc8cced343476dbe03ebe108fd44d85309