Merge aura::DispatcherLinux into base::MessagePumpAuraX11.
The majority of incoming XEvents are related to specific XWindows and need to
be sent to a specific MessagePumpDispatcher. MessagePumpAuraX11 would pass the
XEvent to aura::DispatcherLinux, which would pass it to the next
dispatcher. Because of the chromeos unit tests, which spin up a new XDisplay on
each test, these two objects need to have equivalent lifetimes. Due to the
incoming clipboard support (which this patch was split off from),
DispatcherLinux can't live in aura:: anymore. The solution is to merge the two
objects that have to have equivalent lifetimes together.

The one ui:: specific check was split off into an MessagePumpObserver. This also got rid of those hacks where we would static downcast to DispatcherLinux
from MessagePumpDispatcher.

BUG=130805

Review URL: https://chromiumcodereview.appspot.com/10895020

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@154174 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4168ad750ebe1bf59b29fd82dc9425c016a404e6