events: Inlcude the event-phase and event-result in Event.
An event-handler can be installed on a target as both a pre-target and a post-target handler.
In such cases, it will be necessary to know the phase of the event-dispatch. So add EventPhase.
Also, from a handler, it may be necessary to know whether the event has been handled by any other
handlers before this. So include EventResult in the event too.

BUG=none.

Review URL: https://chromiumcodereview.appspot.com/10916210

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@155825 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: daafa7126ac990796195cd9c2b3d875ac86acadb