Add image resources for "Display" item in the uber tray.
Previously I submitted a CL for this, but it was reverted due to
build break of win_aura.  To prevent the error again, I splitted
it into two, and this is the first CL to add the resource files.

BUG=141394

Review URL: https://chromiumcodereview.appspot.com/10905218

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@156116 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1338ed8c078a686ad542140456d545652af7abff