Make app launcher on Windows pinnable.
This sets the appropriate property store values so that when pinned the created shortcut will have the right icon and launch switches. A .ico has been added to the resources for this.

This also removes the need to set the window icon in the previous way, so the previous resource and code has been removed.

BUG=151672

Review URL: https://chromiumcodereview.appspot.com/10996042

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@159225 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a2ad3a1753ee67c96b6ae5eee9bfd214bd7d1764