Fix wrong truncation of notification popup messages in ash.
views::Label control first calculates the size of the text,
and then tries to render the text in the exact size of rectangle.
If < is used instead of <= in the patched line, the control
flows to the BreakIterator mode, that sometimes computes a
larger width value and causes truncation.

BUG=155663

Review URL: https://codereview.chromium.org/11150013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@162308 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 54db05eab3af2554886bd2fab8783f7909dca6fe