Fixed: on windows, fixed-size app windows would lose their shadows.
Some code checked the resizability of the window and used it to key a change in
window style. For app windows, we always want a very particular window style,
so we check for the remove_standard_frame Widget::InitParams option and ignore
the resize flag.

Also, change the shell window NonClientHitTest code to not return HTTOPLEFT and
friends when the window is non-resizable.

R=ben@chromium.org
BUG=155219

Review URL: https://chromiumcodereview.appspot.com/11154007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@162387 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 585e1331aa544d62587c5f2daf1207300876eff5